requirements.txt
pandas==2.3.2
python_dateutil==2.9.0
